% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2_sDiSCA}
\alias{example2_sDiSCA}
\title{Punctuation used by a selection of French and English writers}
\format{
\subsection{\code{example2_sDiSCA}}{

A data frame
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
example2_sDiSCA
}
\description{
The example for sparse DiCA
}
\keyword{datasets}
