% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example3_sMCA}
\alias{example3_sMCA}
\title{Chinese Version of the Independent and Interdependent Self Scale (C-IISS) Dataset}
\format{
A data frame with 130 rows and 42 columns representing the Likert scale responses
for each item, along with demographic information of the participants.
}
\source{
Lu et al. (2007). Development of the Chinese version of the Independent and Interdependent
Self Scale. Journal reference.
}
\usage{
example3_sMCA
}
\description{
This dataset contains responses from 130 undergraduate students (77 females and 53 males,
mean age = 19.49, SD = 1.52) from National Cheng Kung University to the Chinese version
of the Independent and Interdependent Self Scale (C-IISS). The scale, developed by
Lu et al. (2007), comprises forty-two 7-point Likert scale items (1 = strongly disagree,
7 = strongly agree), with 21 items measuring independence (awareness and value of oneself
as an individual) and 21 measuring interdependence (valuation and actions based on one's
cohort). Participants provided written informed consent and were compensated with NTD 120.
}
\details{
Prior to analysis with sMCA, responses for each item were binned into categories of
comparable sizes to mitigate the influence of event rarity in MCA (and sMCA) results.
The association patterns between these items are illustrated in Supplementary Figures
(refer to Figure S1 for binning and Figure S2 for association patterns).
}
\references{
Lu, L. (2007). Developing the Chinese version of the Independent and Interdependent Self Scale.
Journal Title, Volume(Issue), Page range.
Supplementary Material: Figures S1 (Binning of Responses) and S2 (Association Patterns).
}
\keyword{datasets}
