% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_discriminant}
\alias{is_discriminant}
\title{is_discriminant}
\usage{
is_discriminant(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class discriminant, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "discriminant"
}
\details{
The functions in the \code{sGSVD} package that perform discriminant analysis. The function is \code{\link{sparseDiCA}}.
}
\seealso{
\code{\link{inherits}}
}
