% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_mca}
\alias{is_mca}
\title{is_mca}
\usage{
is_mca(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class mca, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "mca"
}
\details{

}
\seealso{
\code{\link{inherits}}
}
