% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sca}
\alias{is_sca}
\title{is_sca}
\usage{
is_sca(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sca, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sca"
}
\details{
The functions in the \code{sGSVD} package that perform discriminant simple correspondence analysis. The function is \code{\link{sparseDiSCA}}.
}
\seealso{
\code{\link{inherits}}
}
