% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sEIGEN}
\alias{is_sEIGEN}
\title{is_sEIGEN}
\usage{
is_sEIGEN(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sEIGEN, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sEIGEN"
}
\details{
The functions in the \code{sGSVD} package that are analyzed by sparse eigendecomposition \code{sparseEIGEN}. For example, \code{sparseMDS} and \code{sparseDiSTATIS}.
}
