% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sGEIGEN}
\alias{is_sGEIGEN}
\title{is_sGEIGEN}
\usage{
is_sGEIGEN(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sGEIGEN, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sGEIGEN"
}
\details{
The functions in the \code{sGSVD} package that are analyzed by sparse generalized eigendecomposition \code{sparseGEIGEN}. For example, \code{sparseMDS} and \code{sparseDiSTATIS}.
}
