% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sGPCA}
\alias{is_sGPCA}
\title{is_sGPCA}
\usage{
is_sGPCA(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sPCA, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sGPCA"
}
\details{
The three primary functions in the \code{sGPCA} package produce an inherited (hierarchical) class structure where all of them are of type "sGPCA". The function is \code{\link{sparseMFA}}.
}
\seealso{
\code{\link{inherits}}
}
