% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sGSVD}
\alias{is_sGSVD}
\title{is_sGSVD}
\usage{
is_sGSVD(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sGSVD, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sGSVD"
}
\details{
The three primary functions in the \code{sGSVD} package produce an inherited (hierarchical) class structure where all of them are of type "sGSVD". Those functions are \code{\link{sparseGSVD}}, \code{\link{sparseCA}}, \code{\link{sparseMCA}}, \code{\link{sparseMCA}}, \code{\link{sparseMFA}}, \code{\link{sparseDiCA}}, and \code{\link{sPLSCA}}.
}
\seealso{
\code{\link{inherits}}
}
