% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sMDS}
\alias{is_sMDS}
\title{is_sMDS}
\usage{
is_sMDS(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sMDS, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sMDS"
}
\details{
The functions in the \code{sGSVD} package that are analyzed by sparse multidimensional scaling \code{sparseMDS}.
}
