% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sPLS}
\alias{is_sPLS}
\title{is_spls}
\usage{
is_sPLS(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sGSVD, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sGSVD"
}
\details{
The three primary functions in the \code{sGSVD} package produce an inherited (hierarchical) class structure where all of them are of type "spls". Those functions are \code{\link{sparsePLSC}} and \code{\link{sparsePLSCA}}.
}
\seealso{
\code{\link{inherits}}
}
