% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_sSVD}
\alias{is_sSVD}
\title{is_sSVD}
\usage{
is_sSVD(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class sGSVD, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "sSVD"
}
\details{
The three primary functions in the \code{sGSVD} package produce an inherited (hierarchical) class structure where all of them are of type "sSVD". Those functions are \code{\link{sparseSVD}} and \code{\link{sparsePLSC}}.
}
\seealso{
\code{\link{inherits}}
}
