% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_select.R
\name{partition_select}
\alias{partition_select}
\title{Create a partition from a variable}
\usage{
partition_select(vect, dta = NULL, choice = "obs")
}
\arguments{
\item{vect}{the factor vector used to create the partition. The vector length must be equal to the number of observation or of variables.}

\item{dta}{the complete data set used in the case of the variables (columns) partition.}

\item{choice}{Indicate if you want a partition on the observations ("obs") or on the variables ("var") of the data set (default obs).}
}
\value{
Returns a list with the index of each partition levels.
}
\description{
Create a partition from a variable
}
\examples{
data(cheese)
G <- partition_select (cheese [,3])
vect <- sample(c("a", "b"), size = ncol(cheese), replace = T)
G2 <- partition_select (vect, dta = cheese, choice = "var")
}
