% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_variables.R
\name{partition_variables}
\alias{partition_variables}
\title{Create a partition by variable}
\usage{
partition_variables(Y)
}
\arguments{
\item{X}{a qualitative variable data frame or matrix}
}
\value{
Returns the partition by variable of a qualitative data set.
}
\description{
Create a partition by variable
}
\examples{
data(cheese)
partition_variables(cheese)
}
