% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sDIMCA.R
\name{sparseDiMCA}
\alias{sparseDiMCA}
\title{Group sparse Discriminant Correspondence Analysis for MCA}
\usage{
sparseDiMCA(
  DATA,
  design,
  components = 0,
  tol = .Machine$double.eps,
  make_data_nominal = TRUE,
  make_design_nominal = TRUE,
  doublecentering = TRUE,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{epsPOCS}{}
}
\description{
Group sparse Discriminant Correspondence Analysis for MCA
}
