% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sDiSTATIS.R
\name{sparseDiSTATIS}
\alias{sparseDiSTATIS}
\title{Sparse DiSTATIS}
\usage{
sparseDiSTATIS(
  DATA,
  method = "distance",
  Cmat.is.RV = TRUE,
  masses.Cmat = NULL,
  masses.Splus = NULL,
  sparse.Cmat = FALSE,
  sparse.Splus = TRUE,
  components.Cmat = 0,
  components.Splus = 0,
  init.Cmat = "svd",
  init.Splus = "svd",
  seed = NULL,
  grp.Cmat = NULL,
  grp.Splus = NULL,
  rds.Cmat = rep(1, components.Cmat),
  rds.Splus = rep(1, components.Splus),
  orthogonality.Cmat = "loadings",
  OrthSpace.Cmat = NULL,
  orthogonality.Splus = "loadings",
  OrthSpace.Splus = NULL,
  projPriority.Cmat = "orth",
  projPriority.Splus = "orth",
  itermaxALS.Cmat = 1000,
  itermaxPOCS.Cmat = 1000,
  itermaxALS.Splus = 1000,
  itermaxPOCS.Splus = 1000,
  epsALS.Cmat = 1e-10,
  epsPOCS.Cmat = 1e-10,
  epsALS.Splus = 1e-10,
  epsPOCS.Splus = 1e-10,
  tol = .Machine$double.eps
)
}
\arguments{
\item{DATA}{a cube of distance matrix}

\item{method}{"distance" or "covariance"}

\item{masses.Cmat}{a vector of masses for the rows/columns of the rv matric. Default is set to "NULL" with all 1s.}

\item{masses.Splus}{a vector of masses for the rows/columns of each table in the cube. Default is set to "NULL" with equal masses of 1/(the number of rows) being used.}

\item{tol}{}
}
\description{
Sparse DiSTATIS
}
