% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMCA.R
\name{sparseMCA}
\alias{sparseMCA}
\title{Group Sparse Multiple Correspondence Analysis}
\usage{
sparseMCA(
  DATA,
  components = 0,
  tol = .Machine$double.eps,
  make_data_nominal = TRUE,
  doublecentering = TRUE,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  correction4SI = "mca",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{DATA}{The contingency table for analysis.}

\item{components}{The number of dimensions for the analysis; defaults to 0.}

\item{tol}{Tolerance for the convergence criterion; defaults to \code{.Machine$double.eps}.}

\item{make_data_nominal}{Logical indicating whether to treat DATA as nominal (TRUE) or not (FALSE); defaults to TRUE.}

\item{doublecentering}{Logical indicating whether double centering should be applied; defaults to TRUE.}

\item{init}{Method for initialization; defaults to "svd".}

\item{initLeft}{Initial values for the left side; defaults to NULL.}

\item{initRight}{Initial values for the right side; defaults to NULL.}

\item{seed}{Seed for random number generation; defaults to NULL.}

\item{rdsLeft}{Radii for the left side; defaults to rep(1, components).}

\item{rdsRight}{Radii for the right side; defaults to rep(1, components).}

\item{grpLeft}{Grouping vector for the rows; defaults to NULL.}

\item{grpRight}{Grouping vector for the columns; required if DATA is not treated as nominal.}

\item{orthogonality}{Type of orthogonality constraint; defaults to "loadings".}

\item{OrthSpaceLeft}{Orthogonal space for the left side; defaults to NULL.}

\item{OrthSpaceRight}{Orthogonal space for the right side; defaults to NULL.}

\item{projPriority}{Priority of the projection; defaults to "orth".}

\item{projPriorityLeft}{Priority of the left projection; defaults to projPriority.}

\item{projPriorityRight}{Priority of the right projection; defaults to projPriority.}

\item{correction4SI}{Correction method for supplementary items; defaults to "mca".}

\item{itermaxALS}{Maximum number of ALS iterations; defaults to 1000.}

\item{itermaxPOCS}{Maximum number of POCS iterations; defaults to 1000.}

\item{epsALS}{Convergence criterion for ALS; defaults to 1e-10.}

\item{epsPOCS}{Convergence criterion for POCS; defaults to 1e-10.}
}
\value{
Returns an object of class c("sSVD", "sGSVD", "list") containing the results of the sparse multiple correspondence analysis. This object includes the results of sparse generalized singular value decomposition (sparseGSVD) and additional data specific to MCA.
}
\description{
Performs group sparse multiple correspondence analysis (MCA) on a given contingency table.
}
\examples{
# Example usage of sparseMCA function
# Assuming `data` is a contingency table
#' \dontrun{result <- sparseMCA(DATA = data)}
}
