% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMDS.R
\name{sparseMDS}
\alias{sparseMDS}
\title{Sparse Multidimensional Scaling}
\usage{
sparseMDS(
  DATA,
  masses = NULL,
  components = 0,
  tol = .Machine$double.eps,
  method = "distance",
  init = "svd",
  seed = NULL,
  grp = NULL,
  rds = rep(1, components),
  orthogonality = "loadings",
  OrthSpace = NULL,
  projPriority = "orth",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{DATA}{The data matrix for MDS analysis.}

\item{masses}{A vector of masses for the rows/columns; defaults to NULL with equal masses of 1/(number of rows).}

\item{components}{The number of dimensions for the analysis; defaults to 0.}

\item{tol}{Tolerance for the convergence criterion; defaults to \code{.Machine$double.eps}.}

\item{method}{Method for analysis, either "distance" or "covariance".}

\item{init}{Initialization method; defaults to "svd".}

\item{seed}{Seed for random number generation; defaults to NULL.}

\item{rds}{Radii for the analysis; defaults to rep(1, components).}

\item{orthogonality}{Type of orthogonality constraint; defaults to "loadings".}

\item{OrthSpace}{Orthogonal space for the analysis; defaults to NULL.}

\item{projPriority}{Priority of the projection; defaults to "orth".}

\item{itermaxALS}{Maximum number of ALS iterations; defaults to 1000.}

\item{itermaxPOCS}{Maximum number of POCS iterations; defaults to 1000.}

\item{epsALS}{Convergence criterion for ALS; defaults to 1e-10.}

\item{epsPOCS}{Convergence criterion for POCS; defaults to 1e-10.}
}
\value{
Returns an object containing the results of the sparse MDS analysis, including eigenvalues, eigenvectors, and various metrics relevant to the analysis.
}
\description{
Performs sparse multidimensional scaling (MDS) on a given data matrix.
}
\examples{
# Example usage of sparseMDS function
# Assuming `data` is a data matrix
#' \dontrun{result <- sparseMDS(DATA = data)}

}
