% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMFA.R
\name{sparseMFA}
\alias{sparseMFA}
\title{Sparse Multiple Factor Analysis}
\source{
Some lines of the function are inspired by the MExPosition package by Derek Beaton and Cherise Chin Fatt.
}
\usage{
sparseMFA(
  X,
  column.design,
  components = 0,
  sparseOption = "variable",
  center = TRUE,
  scale = TRUE,
  mfa.scale = TRUE,
  tol = .Machine$double.eps,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{Data matrix with I rows and J columns.}

\item{column.design}{Design vector indicating the grouping of columns.}

\item{components}{The number of dimensions to extract; defaults to 0.}

\item{sparseOption}{"variable" or "subtable" indicating the sparsity option.}

\item{center}{Logical or numeric vector for centering each column of X; passed to \code{\link{scale}}.}

\item{scale}{Logical or numeric vector for scaling each column of X; passed to \code{\link{scale}}.}

\item{tol}{Tolerance for the convergence criterion; defaults to \code{.Machine$double.eps}.}

\item{init}{Initialization method; defaults to "svd".}

\item{initLeft}{Initial values for the left side; defaults to NULL.}

\item{initRight}{Initial values for the right side; defaults to NULL.}

\item{seed}{Seed for random number generation; defaults to NULL.}

\item{rdsLeft}{Radii for the left side; defaults to rep(1, components).}

\item{rdsRight}{Radii for the right side; defaults to rep(1, components).}

\item{grpLeft}{Grouping vector for the left side; defaults to NULL.}

\item{grpRight}{Grouping vector for the right side; based on sparseOption.}

\item{orthogonality}{Type of orthogonality constraint; defaults to "loadings".}

\item{OrthSpaceLeft}{Orthogonal space for the left side; defaults to NULL.}

\item{OrthSpaceRight}{Orthogonal space for the right side; defaults to NULL.}

\item{projPriority}{Priority of the projection; defaults to "orth".}

\item{itermaxALS}{Maximum number of ALS iterations; defaults to 1000.}

\item{itermaxPOCS}{Maximum number of POCS iterations; defaults to 1000.}

\item{epsALS}{Convergence criterion for ALS; defaults to 1e-10.}

\item{epsPOCS}{Convergence criterion for POCS; defaults to 1e-10.}
}
\value{
Returns an object containing the results of the sparse MFA analysis. This object includes the results of sparse generalized singular value decomposition (sparseGSVD) and additional data specific to MFA.
}
\description{
Performs sparse multiple factor analysis (MFA) on a given data matrix.
}
\examples{
# Example usage of sparseMFA function
# Assuming `data` is a data matrix and `column.design` is the design vector
\dontrun{result <- sparseMFA(X = data, column.design = column.design)}

}
