% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLSC.R
\name{sparsePLSC}
\alias{sparsePLSC}
\title{Sparse Partial Least Square Correlation}
\source{
Some arguments of the function are inspired by the GPLS package by Derek Beaton.
}
\usage{
sparsePLSC(
  X,
  Y,
  components = 0,
  center_X = TRUE,
  center_Y = TRUE,
  scale_X = TRUE,
  scale_Y = TRUE,
  tol = .Machine$double.eps,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "both",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{Data matrix with I rows and J columns.}

\item{Y}{Data matrix with I rows and K columns.}

\item{components}{The number of dimensions to extract; defaults to 0.}

\item{center_X}{Logical or numeric vector for centering each column of X; passed to \code{\link{scale}}.}

\item{center_Y}{Logical or numeric vector for centering each column of Y; passed to \code{\link{scale}}.}

\item{scale_X}{Logical or numeric vector for scaling each column of X; passed to \code{\link{scale}}.}

\item{scale_Y}{Logical or numeric vector for scaling each column of Y; passed to \code{\link{scale}}.}

\item{tol}{Tolerance for the convergence criterion; defaults to \code{.Machine$double.eps}.}

\item{init}{Initialization method; defaults to "svd".}

\item{initLeft}{Initial values for the left side; defaults to NULL.}

\item{initRight}{Initial values for the right side; defaults to NULL.}

\item{seed}{Seed for random number generation; defaults to NULL.}

\item{rdsLeft}{Radii for the left side; defaults to rep(1, components).}

\item{rdsRight}{Radii for the right side; defaults to rep(1, components).}

\item{grpLeft}{Grouping vector for the left side; defaults to NULL.}

\item{grpRight}{Grouping vector for the right side; defaults to NULL.}

\item{orthogonality}{Type of orthogonality constraint; defaults to "both".}

\item{OrthSpaceLeft}{Orthogonal space for the left side; defaults to NULL.}

\item{OrthSpaceRight}{Orthogonal space for the right side; defaults to NULL.}

\item{projPriority}{Priority of the projection; defaults to "orth".}

\item{projPriorityLeft}{Priority of the left projection; defaults to projPriority.}

\item{projPriorityRight}{Priority of the right projection; defaults to projPriority.}

\item{itermaxALS}{Maximum number of ALS iterations; defaults to 1000.}

\item{itermaxPOCS}{Maximum number of POCS iterations; defaults to 1000.}

\item{epsALS}{Convergence criterion for ALS; defaults to 1e-10.}

\item{epsPOCS}{Convergence criterion for POCS; defaults to 1e-10.}
}
\value{
Returns an object containing the results of the sparse partial least square correlation analysis. This object includes results from the sparse generalized singular value decomposition (sparseGSVD) applied to X and Y, and additional data specific to the analysis.
}
\description{
Performs sparse partial least square correlation analysis on two data matrices.
}
\examples{
# Example usage of sparsePLSC function
# Assuming `X` and `Y` are data matrices
\dontrun{result <- sparsePLSC(X = X, Y = Y)}

}
