% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sSTATIS.R
\name{sparseSTATIS}
\alias{sparseSTATIS}
\title{sparse STATIS}
\source{
Some lines of the function are inspired by the MExPosition package by Derek Beaton and Cherise Chin Fatt.
}
\usage{
sparseSTATIS(
  X,
  column.design,
  mfa.normalize = TRUE,
  sparse.Cmat = FALSE,
  sparse.grandX = TRUE,
  components.Cmat = 0,
  components.grandX = 0,
  sparseOption = "variable",
  center = TRUE,
  scale = TRUE,
  tol = .Machine$double.eps,
  init.Cmat = "eig",
  init.grandX = "svd",
  initLeft.grandX = NULL,
  initRight.grandX = NULL,
  seed = NULL,
  rdsLeft.Cmat = rep(1, components.Cmat),
  rdsRight.Cmat = rep(1, components.Cmat),
  rdsLeft.grandX = rep(1, components.grandX),
  rdsRight.grandX = rep(1, components.grandX),
  grp.Cmat = NULL,
  grpLeft.grandX = NULL,
  grpRight.grandX = NULL,
  orthogonality.Cmat = "loadings",
  OrthSpaceLeft.Cmat = NULL,
  OrthSpaceRight.Cmat = NULL,
  orthogonality.grandX = "loadings",
  OrthSpaceLeft.grandX = NULL,
  OrthSpaceRight.grandX = NULL,
  projPriority.Cmat = "orth",
  projPriorityLeft.Cmat = projPriority.Cmat,
  projPriorityRight.Cmat = projPriority.Cmat,
  projPriority.grandX = "orth",
  projPriorityLeft.grandX = projPriority.grandX,
  projPriorityRight.Slus = projPriority.grandX,
  itermaxALS.Cmat = 1000,
  itermaxPOCS.Cmat = 1000,
  itermaxALS.grandX = 1000,
  itermaxPOCS.grandX = 1000,
  epsALS.Cmat = 1e-10,
  epsPOCS.Cmat = 1e-10,
  epsALS.grandX = 1e-10,
  epsPOCS.grandX = 1e-10
)
}
\arguments{
\item{X}{Data matrix with \emph{I} rows and \emph{J} columns}

\item{mfa.normalize}{TRUE (default) or FALSE.}

\item{sparse.Cmat}{TRUE or FALSE (default). If the C matrix (e.g., the RV matrix) is sparsified to obtain the weights.}

\item{components.Cmat}{the number of dimensions to extract from the C matrix (e.g., the RV matrix).}

\item{sparseOption}{"variable" or "subtable"}

\item{center}{For the \code{X} matrix: A parameter to pass through to \code{center} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{X}.}

\item{scale}{For the \code{X} matrix: A parameter to pass through to \code{scale} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{X}.}

\item{tol}{default is .Machine$double.eps. A parameter to pass through to \code{\link[GSVD]{gplssvd}}; eliminates singular values that are effectively zero (and thus drops null components).}

\item{sparse.Splus}{TRUE (default) or FALSE. If the compromise is sparsified.}

\item{components.Splus}{the number of dimensions to extract from the compromise.}

\item{epsPOCS.Splus}{}
}
\description{
sparse STATIS
}
