% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_disjonctif.R
\name{tab_disjonctif}
\alias{tab_disjonctif}
\title{Disjonctif table}
\usage{
tab_disjonctif(X)
}
\arguments{
\item{X}{A data frame or matrix with factors}
}
\value{
Returns the disjonctif table
}
\description{
Disjonctif table
}
\examples{
library(FactoMineR)
data(tea)
tea <- tea [, 1:8]

tab_disjonctif(tea)

}
