% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_sCA}
\alias{example1_sCA}
\title{Cause of death data}
\format{
\subsection{\code{example1_sCA}}{

A data frame
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
example1_sCA
}
\description{
The example for sparse CA
}
\keyword{datasets}
