% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2_sDiSCA}
\alias{example2_sDiSCA}
\title{Punctuation Marks in Literary Works from Project Gutenberg}
\format{
A data frame where each row represents a literary work and includes
the following columns for punctuation mark counts: \code{Comma}, \code{Period}, \code{QuestionMark},
\code{ExclamationMark}, \code{Colon}, \code{Semicolon}, \code{Apostrophe}, \code{EnglishQuotes}, \code{FrenchQuotes},
\code{Dash}, \code{MDash}, along with metadata about the work such as \code{Author}, \code{Country}, and \code{Period}.

\subsection{\code{example2_sDiSCA}}{

A data frame
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
Project Gutenberg, processed using the \code{gutenbergr} R package.

https://www.gutenberg.org/
}
\usage{
example2_sDiSCA
}
\description{
This dataset compiles the counts of punctuation marks in literary works
by authors of different times and origins. The texts were sourced from
the Gutenberg Project and processed using the R \code{gutenbergr} package.
The dataset includes authors from France (N = 63), the United Kingdom (N = 61),
and the United States (N = 36) spanning various periods (refer to Table 1 for details).
}
\details{
The dataset focuses on the frequency of nine punctuation marks: comma (,),
period (.), question mark (?), exclamation mark (!), colon (:), semicolon (;),
apostrophe ('), English quotation marks (“ ”), French quotation marks (« »),
dashes (-), and M-dashes (—). Works that were translations were excluded to
ensure the dataset only contains books written in the authors' original languages.
}
\references{
Robinson, D. (2021). gutenbergr: Download and Process Public Domain Works from Project Gutenberg. R package version 0.2.0.
\url{https://CRAN.R-project.org/package=gutenbergr}
}
\keyword{datasets}
