% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example3_sMCA}
\alias{example3_sMCA}
\title{Questionnaire Data}
\format{
\subsection{\code{example3_sMCA}}{

A data frame
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
example3_sMCA
}
\description{
The example for sparse MCA
}
\keyword{datasets}
