% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example4_sDiMCA}
\alias{example4_sDiMCA}
\title{Math quiz}
\format{
\subsection{\code{example4_sDiMCA}}{

A data frame
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
example4_sDiMCA
}
\description{
The example for sparse DiMCA
}
\keyword{datasets}
