% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_mca}
\alias{is_mca}
\title{Check if an Object is of Class "MCA"}
\usage{
is_mca(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
A logical value: \code{TRUE} if the object is of class "MCA", \code{FALSE} otherwise.
}
\description{
Tests if the provided object \code{x} is of class type "MCA".
}
\details{
This function is useful for validating whether an object belongs to the class "MCA",
which is relevant for functions in the \code{sGSVD} package, particularly for those performing
discriminant multiple correspondence analysis, such as \code{sparseDiMCA}.
}
\seealso{
\code{\link{inherits}} for checking object inheritance.
}
