% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCA.R
\name{sparseCA}
\alias{sparseCA}
\title{Group sparse Correspondence Analysis}
\usage{
sparseCA(
  DATA,
  components = 0,
  tol = .Machine$double.eps,
  doublecentering = TRUE,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{DATA}{the contingency table}

\item{epsPOCS}{}

\item{k}{the number of dimensions}
}
\description{
Group sparse Correspondence Analysis
}
