% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMCA.R
\name{sparseMCA}
\alias{sparseMCA}
\title{Group sparse Multiple Correspondence Analysis}
\usage{
sparseMCA(
  DATA,
  components = 0,
  tol = .Machine$double.eps,
  make_data_nominal = TRUE,
  doublecentering = TRUE,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  correction4SI = "mca",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{DATA}{the contingency table}

\item{grpLeft}{the grouping vector for the rows}

\item{grpRight}{the grouping vector for the columns}

\item{correction4SI}{}
}
\description{
Group sparse Multiple Correspondence Analysis
}
