% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMDS.R
\name{sparseMDS}
\alias{sparseMDS}
\title{Sparse Multidimensional Scaling}
\usage{
sparseMDS(
  DATA,
  masses = NULL,
  components = 0,
  tol = .Machine$double.eps,
  method = "distance",
  init = "svd",
  seed = NULL,
  grp = NULL,
  rds = rep(1, components),
  orthogonality = "loadings",
  OrthSpace = NULL,
  projPriority = "orth",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{masses}{a vector of masses for the rows/columns. Default is set to "NULL" with equal masses of 1/(the number of rows) being used.}

\item{method}{"distance" or "covariance"}

\item{OrthSpace}{}
}
\description{
Sparse Multidimensional Scaling
}
