% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMFA.R
\name{sparseMFA}
\alias{sparseMFA}
\title{sparse Multiple Factor Analysis}
\source{
Some lines of the function are inspired by the MExPosition package by Derek Beaton and Cherise Chin Fatt.
}
\usage{
sparseMFA(
  X,
  column.design,
  components = 0,
  sparseOption = "variable",
  center = TRUE,
  scale = TRUE,
  mfa.scale = TRUE,
  tol = .Machine$double.eps,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{Data matrix with \emph{I} rows and \emph{J} columns}

\item{components}{the number of dimensions to extract}

\item{sparseOption}{"variable" or "subtable"}

\item{center}{For the \code{X} matrix: A parameter to pass through to \code{center} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{X}.}

\item{scale}{For the \code{X} matrix: A parameter to pass through to \code{scale} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{X}.}

\item{tol}{default is .Machine$double.eps. A parameter to pass through to \code{\link[GSVD]{gplssvd}}; eliminates singular values that are effectively zero (and thus drops null components).}

\item{epsPOCS}{}
}
\description{
sparse Multiple Factor Analysis
}
