% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLSCA.R
\name{sparsePLSCA}
\alias{sparsePLSCA}
\title{sparse Partial Least Square Correspondence anlaysis}
\usage{
sparsePLSCA(
  X,
  Y,
  Mx = NULL,
  My = NULL,
  Wx = NULL,
  Wy = NULL,
  components = 2L,
  tol = .Machine$double.eps,
  doublecentering_X = TRUE,
  doublecentering_Y = TRUE,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "both",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{epsPOCS}{}

\item{DATA}{the contingency table}

\item{k}{the number of dimensions}
}
\description{
sparse Partial Least Square Correspondence anlaysis
}
